April 20 changes
Amendment 1: Addition of "The Vice President will only vote in the President election in the event of a tie during the runoff or a tie in the first round that would result in more than two candidates making the runoff." to Article VII Section E. This will prevent the presidential candidate from breaking their own tie, in the event that a tie were to happen in the presidential election. 

Amendment 2: Ranked voting. Our new voting procedure will follow ranked-voting (AKA instant run-off) procedure. This makes absentee ballots easier to manage. It also enables people to vote for the candidate they believe will best fill the role, not the one who is most likely to beat the candidate they dislike.  

Amendment 3: Active membership clarifications. Sign-in is now required.

Amendment 4: Absentee ballots. 

Amendment 5: You do not have to be an active member to run for an officer position.

Failed amendment: Quorum. We voted on having a minimum percentage of the club present at elections. This vote did not pass.